Adding suspense to add a loading animation when loading blogs
