+++ bootcamp-milestone-4/components/centeredContainer/centeredContainer.module.css
@@ -1,0 +1,8 @@
+.centeredContainer {
+    position: absolute;
+    display: flex;
+    width: 100%;
+    height: 100%;
+    justify-content: center; /* horizontal */
+    align-items: center;     /* vertical */
+}
